Do some tricks to get sssd behave nicely with shared laptops where a new user is using the laptop right after it connects to the network. LDAP provider in sssd 1.9.4 takes some time to realise that it can go online and during that time the user has already tried to login multiple times. Here we work around that by not using ldap id_provider in sssd, but use the sssd proxy provider with extrausers. This seems to fix the problems for now and also give some extra flexibility in DNS SRV record handling.
If pam_krb5 succeeds, jump to fetch extrausers info with
puavo-ltsp-login. In that case pam_sss is optional as we already have a
working kerberos ticket, so sssd malfunction can be ignored.

If pam_krb5 fails, pam_sss needs to succeed for the user to be able to
login. In that case do not call puavo-ltsp-login as it could not connect
a working server.