bootserver: fix nbd-server SIGHUPing
Previously, handle_imagefiles() returned non-zero if cmp failed (for
example when $conf did not exist because the image file was new) which
caused the script to exit (because set -e).

However, the whole cmp thing was unnecessary because nbd-server does it
internally anyways. We can just write the configuration for each image
file and then signal the running nbd-server and let it do the rest.