Revert "client: No need to create session dir with mkdir -p as the parent directory"
Actually, we need to use -p, because the ${PAM_USER} dir might exist in
which case the mkdir without -p would fail and the following chmod would
not be executed and the dir might not be 0700. Of course we could use
conditionals or set +e or something like that, but I think the cleanest
way to do it is to assume that the dir might already exists (-p).

It is assumed that /var/lib/puavo-desktop/users exists always, it is
created during installation. If it does not exists,
/var/lib/puavo-desktop/users (or any preceding missing component) might
get created with wrong access modifiers resulting in totally broken
session caching. But defending against broken installation is not
sustainable, the installation might be broken in million different ways
too.

This reverts commit af48586332bee01ca7d9017d93bdd9bdde22568e.