Rewrite in Python3
Bash / POSIX Shell was getting way too limited for this kind of work. I
was mainly missing proper data structures.

lib/_dcf.py is a good start for a quite generic Debian control file
parser. If it evolves enough, it would be wise to extract it to an
independent library.