Set default kernel to 3.12.10.opinsys1
3.13.1 causes a really strange nbd-connection problem:

If there is a hanging nbd-connection (nbd-server process still exists
but the client has terminated), the client host to which the connection
is associated cannot make a new connection until the nbd-server has
died.

This does not seem to happen with 3.12.10.opinsys1.

Moreover, all zram patches included in 3.13.1 are also included in
3.12.10.