ap: ignore SIGCHLD during exit
Previously, when SIGCHLD trap was set, pkill caused on_sigchld to
execute in the middle of on_exit. on_sigchld in turn calls exit aborts
the cleanup process. This commit minimizes the risk of exiting during
on_exit by ignoring SIGCHLD. Note however, that there is still a race
condition between the on_exit and SIGCHLD of critical childs, but it is
unavoidable because Bash does not have atomic signal masking
capabilities.