client: add support for persistent image overlays
This, a so called persisent image overlay (aka. "superlaptop") mode, is
activated when the device boots from a local image and a kernel argument
puavo.image.overlay=NAME is defined.

A directory /images/overlays/IMAGE/NAME is created if it does not exist
and then mounted as a writable layer on top of the filesystem from the
image. IMAGE is the filename of the image and NAME is an arbitrary name
for the overlay.

Both OverlayFS and AUFS are supported as underlying union filesystems.

This commit changes the mount order slightly: now
/dev/mapper/puavo-images is mounted in init-bottom before init-puavo is
run.