client: ignore existing partitions if 'whole' partition scheme is used
Previously, the partitioner was happy to overwrite existing Linux
partitions but not for example Windows partitions. This made the
installation unnecessarily laborious for devices with pre-installed
Windows.

This commit makes the partitioner respect the will of the user: if the
user has chosen 'whole' then the partitioner respects the choice and
trusts the user knows what she is doing.