Do not use loop= kernel parameter to figure out PUAVO_IMAGE at scripts/init-bottom/puavo_ltsp.  There are two problems with this: on laptops there are /ltsp.img and /ltsp-backup.img aliases, which might both map to different images.  However, persistent image overlays must use different directories for different images, because the overlays are image dependent.  Thus we ask the image itself what it is.
The other problem with using "loop=" is that it only works on laptops.
Not using "loop=" possibly makes the persistent image overlays
possible on fatclients, ltsp servers and such.

This of course creates a dependency with a particular way of building
the image (/etc/ltsp/this_ltspimage_name must exist and be sensible).
If that is broken, though, that should only break the persistent
image overlay handling.