client/restricted-packages: make install and uninstall idempotent operations
It was really hard to track the ACTIVE status of packages hence it is
removed and install and uninstall made idempotent. ACTIVE state
tracking was hard because the underlying filesystem might be
non-persistent.

puavo-restricted-package-tool recorded the package activity status to
/state/restricted-packages which is always persistent. However, all
activation changes made elsewhere did not persist over reboots and
hence the active status reporting was broken in every fresh boot.

Making install and uninstall operations idempotent circumvents the
issue nicely. Drawbacks are that installing a restricted package might
override files in the underlying filesystem and INSTALLED state does
not tell if the package has been properly configured: you always need
to run puavo-restricted-package-tool install PACKAGE to ensure it is
properly installed.