install: return or raise the output of update commands
Update scripts do not exit with an error value if the fail, they just
print error messages. This commit make puavo-client-daemon to report all
the output back to the caller.

BUT, all update scripts should be fixed to fail if they fail. For
example puavo-install-and-update-ltspimages does not fail if it fails to
fetch files..