Enable keyring daemon, but disbale keyring prompts
Keyring daemon is required for nm-applet to work correctly. nm-applet
uses keyring daemon to ask and save network passwords. If keyring daemon
services are disabled, clicking a connection from the menu does not do
anything. When keyring daemon is enabled, a prompt dialog is displayed
for user to enter the password. By default, the keyring daemon also
saves passwords, but we have disabled this feature in our network
manager: see opinsys-debs for commit
525a947c0d7cdabf05b8aebcf30bde635973a2af.

However, when keyring daemons are enabled, a default keyring is opened
if it exists or created if it does not. This causes several prompt
dialogs to be shown to the user. We do not want this because it confuses
our users. This commit prevents prompt dialogs from popping up by
removing their dbus services files.