client/restricted-packages: store all packaging stuff in PRP directory
Now packaging files and upstream files live in separate spaces which
ensures we dont get any filepath clashes.

Because restricted package versions are bound to the version of the
puavo-ltsp-client (because restricted package information is part of
the puavo-ltsp-client package), we need to copy all packaging stuff
from the the template diretory to the package directory to avoid nasty
and subtle version incompatibilities:

All package-specific quirks and installation scripts are specific to a
certain package version which is identified by its upstream pack
checksum. When we want to upgrade the package, we update its checksum
and possibly postinst/prerm scripts, because upstream might have done
something very differently in the new version (for example change the
installation layout or rename files).

Then we make a new version of puavo-ltsp-client with updated packaging
scripts. Now, if we had already installed the previous version of the
package (with the previous version of puavo-ltsp-client), but did not
store its packaging scripts, we would need to use the packaging
scripts from the new puavo-ltsp-client. But then we are in trouble,
because the updated packaging scripts were written for the new
version, and they might not be compatible with the already installed
older package.

To avoid this kind of hassle, we just copy all packaging scripts and
files from the template directory into the package directory when the
package is installed for the first time. Packaging scripts are package
*instance* specific, not package *category* specific.

Long story, hard to explain, read it again and hopefully you get it.