client: make lightdm screen script run in a read loop forever
This commit fixes a quite unbelievable bug: previously, in our Precise
systems, pressing key 9 from the keyboard while in Xorg (running in vt7)
caused the system to reboot. Key 8 did not cause reboot.

The cause for mysterious reboots was that the read command in
/usr/share/ltsp/screen.d/lightdm caught input and the evaluation
proceeded to reboot command.

The controlling tty for /usr/share/ltsp/screen.d/lightdm is tty7. So it
seems that lightdm leaked key 9 event to tty7. But the root cause for
the leak is mystery. It does not happen in Quantal.