Set dconf-setting org/gnome/desktop/background/draw-background to "false".
The rationale for this is not obvious.  With the default value "true",
nautilus crashes when desktop resolution changes, but this occurs
only every third time (that is, resolution must change three times
for the crash to occur).  See bug report
https://github.com/opinsys/puavo/issues/156.

According to the gnome dconf schema, this setting should not have
an effect on anything:

  Have GNOME draw the desktop background.  DEPRECATED: This key is
  deprecated and ignored.

And it is true that backgrounds are drawn as usual (apparently that
is handled gnome-settings-daemon or some such), so there are currently
no known regressions due to this change.  But apparently the
Ubuntu Trusty version of Nautilus still reacts to this setting,
because it crashes when it is set to "true" (and not when it is
"false").