Revert "Merge branch 'master' into trusty"
This reverts commit 8cf6b4a0eb1a422b0f8883c8fe00f7f542c221c4, reversing
changes made to 73fd6870cf05aa58cb5c2dd190421af72db35a89.

The transition from "development in trusty-branch" to "development on
master-branch" caused a problem in this repository, because the old
master contained malfunctional VPN configurations. It has never been
used in production. The old trusty-branch was used in production.