Fix i18n objects in web menu items
There was a bug in the menu item id generation when the web item name
was an i18n object. The generated id became always "[object Object]"
which caused the colliding items to disappear from the allItems
collection since items with the same id are considered to be the same.
This also broke the breadcrumbs navigation.

Now the id is generated by using every key and value from the menu
objects. Collisions should be impossible now.