Rename greeter_firstlogin_background tag to greeter_background_firstlogin.
The rationale for this is not at all obvious, so I'll explain: we want
to update a particular image path, depending on systemimage version
(the old systemimages should use a particular path, and new systemimages
another path).  This is because the path was chosen badly: the image
ended up being in random image rotation where it should not have been.