parts/rules: add ttf-ubuntu-font-family package
Yes, it needs that the package itself is provided to image build
system via some means, local package repository is the only viable
solution right now.

I would like to make Puavo OS *completely* self-contained, and for
example in this case, it would require us to have sources of external
third-party packages included in the source tree (for example in
ports/) and built alongside with image.

Closes #13