parts/core: initramfs: ignore all secondary puavo partitions
init-puavo already takes care of mounting all those partitions if
needed, initramfs init should only take care of mounting the root
filesystem and then chainload to init-puavo. It is much easier to
develop and debug stuff on the rootfs image than on the initramfs.

The only partitions requried to mount the rootfs are puavo-images and
puavo-imageoverlays.