parts/conf: rename puavo-conf-mkdb => puavo-conf-init
We are not necessarily creating a database, we are just initializing
the configuration system. Database is just an implementation detail,
which is hidden behind the library. The backend might be a database,
or it might be a volatile process memory or anything else. The point
is, we do not care.