Remove puavo.hosttype configuration parameter
I'm trying to get rid of the whole hosttype-mess. It will be replaced
by configuration profiles, which define a set of parameter overwrites.

Because configuration profiles work on a different level than
configuration parameters, the profile won't be configurable via the
configuration system. Currently, the only way to select the profile is
via kernel command line.

To keep the system backwards compatible for now, puavo.hosttype boot
parameter is used to define the configuration profile. All references
to hosttype must be replaced by specific configuration parameters and
then profile-wide defaults can be defined.

As a side note, parameter values define how the system should be
configured, they do not define the state of the system. When assigning
a value to a parameter, it might or might not change the actual state
of the system.