parts/rules: remove motd Puppet module
Debian does not support /etc/update-motd.d and by default, displays
just /etc/motd, which is a static file.

To preserve the old behavior, one would need to generate
/run/motd.dynamic which is the displayed by pam_motd.so after
successful SSH logins and optionally modify /etc/motd.

But I'm not completely sure if dynamic motd is needed and if it is,
what information should be displayed. Until there is a real need, it
will be absent.

All in all, this old motd module would need to be completely
rewritten, that's why it's now removed.