parts/rules: force libssl1.0.0 and openssl to a specific version
It's ugly, very ugly, bloody ugly. But with Jessie's libssl1.0.0,
registration does not work, because CSR sent by the client generates
the following error at the CA server:

  OpenSSL::X509::RequestError (nested asn1 error):
  app/models/certificate.rb:12:in `initialize'
  app/models/certificate.rb:12:in `new'
  app/models/certificate.rb:12:in `sign_certificate'
  app/controllers/certificates_controller.rb:68:in `block in create'
  app/controllers/certificates_controller.rb:67:in `create'

And hence it fails to create and return a signed certificate.

If libssl1.0.0 is downgraded to 1.0.1e-2+deb7u20, registration works
as expected.

Closes #45