debs: provide staged builds
Some of our packages are build-depencies of other packages, so we
cannot build them independently. First we need to build packages which
are needed to build other packages, then update the repository and
then install those packages and then finally build the rest of the
packages.

Here, we introduce package build stages. Stage1 includes all packages
which do not build-depend on any Puavo OS package but which are
build-dependencies of other Puavo OS packages. Stage2 includes all the
rest.

To build everything properly, one runs:

  make install-build-deps-stage1
  make stage1
  sudo apt-get update
  make install-build-deps-stage2
  make stage2