manager: force fix the encoding of all toxic string to UTF-8
All toxic strings come from LDAP and all of them are UTF-8 encoded. However,
Ruby 1.9 seems to think that they are ASCII-8BIT encoded, which is not the case:
Ruby 1.9 thinks that "Yll\xC3\xA4pito" is ASCII-8BIT encoded string, but it is
actually UTF-8 encoded multibyte representation of "Ylläpito", containing
Unicode character \xe4.

So here we just fix the encoding to UTF-8, so that UTF-8 encoded regexp can be
used to match the string.