Change the disk image compression algorithm to "lzo".
I have tested with "gzip", "lzo", "xz" and no-compression (-noD -noF
-noI -noX), with three hosts, and the boot times were:

            HP (laptop)     Dell (laptop)   Lenovo (fatclient)
  gzip      0m47s           2m33s           1m11s
  lzo       0m44s           1m15s           1m11s
  nocomp    1m5s            1m14s           1m10s
  xz        0m51s           3m14s           1m21s

Here HP = HP Elitebook 2570p, Dell = Dell Latitude 3180, and Lenovo =
Lenovo Thinkcentre Edge.  The laptops' boot times are from pressing Enter
in Grub-menu to active login screen, and the fatclient boot times are
from pressing power button to active login screen.

As can be seen, "lzo" mostly wins on boot time to other alternatives.
The disk image sizes were (otherwise identical images with different
compression algorithms used for squashfs):

  gzip       8824MB
  lzo        9382MB
  nocomp    18561MB
  xz         7927MB

Effects of compression algorithm were also tested on rdiff size.
With a sample size of one, the difference was negligible between
"gzip" and "lzo":

  gzip rdiff size: 1143MB
  lzo  rdiff size: 1118MB

What is curious here is the dramatic speed difference in the boot times
of Dell Latitude 3180 between the "gzip" and "lzo" options.  What is
apparently happening is that as the computer is starting up but before
the kernel and initrd are loaded, it is in a very slow CPU state where
decompression of initrd through "gzip" is a slow operation, taking over
a minute.

These experiences and experiments suggests "lzo" is a better option.
We'll see.  Closes #177.