Filter out duplicate usernames (loginnames) and groupnames in puavo-login.  Previously puavo-login would only guarantee, that there is no other user with the same uid, and no other group with the gid.  Now, it is also guaranteed that there are no other user with the same username/loginname, and no other group with the same groupname.
This is necessary, because there might have been a different user
with the same username previously, but with a different uid.
Home directory path and permissions are unchanged and may thus conflict,
so we can not yet support changing username.  Home directory path
should be tied to a non-changing user id, somehow, some day...