Add http://archive.opinsys.fi/puavo to apt-sources.
Perhaps we should have some other address for this some day.  The
reason for adding this archive is that we should do 32-bit wine-builds
somewhere, and 64-bit environment is not a friendly place to do 32-bit
builds.  So we do the wine build elsewhere and it is available from
http://archive.opinsys.fi/puavo.  The code for doing the build should
be in this repository, though.  Do assign priority of 400 to packages
coming from the remote puavo-os repository, so that packages from there
can not mess up things unexpectedly.