Revert "Update parts/client/Gemfile.lock to get the latest gems"
This reverts commit 161efad1bdd73ddaba0fcf3ee017aa9fe263408c.

It turned out that updating gems slightly broke puavo-client-daemon.
The device image reporting to Puavo got broken, so reverting...
I would rather like to see the bundling removed, and actually
from Debian Stretch all ruby gems can be found with the exception
of "minitest-stub_any_instance", which is needed for tests.
But not doing that now.