Hook homedir cleanup to logins, and do not do cleanup to user that is currently logging in.
This is to prevent accidents in the use case, where the computer is
mostly used by a single person.  We do not want to clean up his/her home
directory at boot time, it might actually be the only home directory
in the computer.  In this case the user can clean up his/her own files
to make some space.  The warning about automatic cleanups is displayed
anyway.