Use 32-bit version of Firefox.
Startup times are maybe a bit longer, something like 30% longer
than the 64-bit version, most probably because we do not have matching
32-bit libraries loaded by other software.  And on a base level
it uses ~60Mb more of memory, again probably for the 32-bit reason.

BUT!  It appears that the 32-bit version uses about 20% less memory
than the 64-bit version for web page content.  If a user has
maybe 12 tabs open, that can mean something like 350Mb difference
in memory consumption between 32-bit and 64-bit.
The memory consumption is much better than on chrome/chromium
(on which there is not much difference between 32-bit and 64-bit),
and it would be nice for us to have some decent browser for low
memory machines as well.

We still need 32-bit Flash for this, though.