Use nfsvers=4.0 for netboot device nfs-mounts (/home).
We used to have this on Trusty and before with mount option
"minorversion=0", but that mount option is broken on Debian Stretch
and I removed it, falling back to NFS version 4.1.  Now we have
encountered some issues with NFS, which may or may be related to
NFS version 4.1, so, we try with 4.0 for now (very quick testing shows
that at least basic functionality works).

Our NFS servers are mostly Linux 3.13 currently, that might also be
relevant.

When process is waiting for an NFS operation that never completes
(it hangs), process stack trace looks like this (process is gnome-shell
in this case):

  root@XXXXXXXXXXX:/proc/4021$ cat /proc/4021/stack
  [<ffffffffc08a3370>] __rpc_wait_for_completion_task+0x30/0x30 [sunrpc]
  [<ffffffffc08a338e>] rpc_wait_bit_killable+0x1e/0x90 [sunrpc]
  [<ffffffffbcae64d7>] mod_timer+0x177/0x3b0
  [<ffffffffc08a3370>] __rpc_wait_for_completion_task+0x30/0x30 [sunrpc]
  [<ffffffffbcab9cf0>] wake_atomic_t_function+0x60/0x60
  [<ffffffffc08a3790>] __rpc_sleep_on_priority+0x340/0x340 [sunrpc]
  [<ffffffffc08a419f>] __rpc_execute+0x14f/0x420 [sunrpc]
  [<ffffffffc0897d43>] rpc_run_task+0xe3/0x120 [sunrpc]
  [<ffffffffc0c49854>] nfs4_call_sync_sequence+0x64/0xa0 [nfsv4]
  [<ffffffffc0c54f69>] nfs4_do_setattr+0x219/0x450 [nfsv4]
  [<ffffffffc0c5524c>] nfs4_proc_setattr+0xac/0x140 [nfsv4]
  [<ffffffffc0bfc92f>] nfs_setattr+0xef/0x210 [nfs]
  [<ffffffffbcc24232>] notify_change+0x2f2/0x460
  [<ffffffffbcc029d5>] chown_common+0x165/0x1c0
  [<ffffffffbcc03ce7>] SyS_fchown+0x87/0xa0
  [<ffffffffbca03b7f>] do_syscall_64+0x8f/0xf0
  [<ffffffffbd0113b8>] entry_SYSCALL_64_after_swapgs+0x42/0xb0
  [<ffffffffffffffff>] 0xffffffffffffffff