Completely overhaul the login screen host info shell extension.
There are too many changes to try to add them in small pieces. The
collector service is run on-demand, so that alone requires numerous
changes to the extension. I'm also adding lots of documentation
about the UI internals, for future reference. There's now a button
that calls the background info collector DBus service, the results
are read asynchronously. The button can be clicked again to update
the info (IP addresses can change, for example) if needed.