Add an ugly fix to mutter for some nasty gnome-shell UI hangs.
Sometimes on our system gnome-shell hangs so that only mouse can
be moved and keyboard remains functional in the focused window,
but otherwise UI elements can not be clicked.  According to logs
produced by

  xdotool key XF86LogGrabInfo

situation is like this:

[   199.122] Active grab 0x40200591 (xi2) on device 'Virtual core pointer' (2):
[   199.122]       client pid 3100 /usr/bin/gnome-shell.distrib
[   199.122]       at 78892 (from passive grab) (device frozen, state 6)
[   199.122]       xi2 event mask for device 2: 0x70000
[   199.122]       passive grab type 4, detail 0x1, activating key 0
[   199.122]       owner-events false, kb 1 ptr 0, confine 0, cursor 0x0
[   199.122] (II) End list of active device grabs

(Event mask is elsewhere logged as "0x70 0 0 0", I take that to mean
mouse button press, button release and motion).
When this happens it can be resolved with (if it is only possible
to type this as user with DISPLAY correctly set):

  setxkbmap -option grab:break_actions
  xdotool key XF86Ungrab

On an automated system with stress testing autopilot, with netbooting
hosts and NFSv4 home directories (may or may not be relevant, probably
not), maybe one in twenty hosts catch the bug in about five minutes.
The problem does not occur if the testpilot user does not click on
anything.

We suspected this was due to something we do (with our patches to
gnome-shell and mutter, plus dconf-settings), but no, it could be
reproduced with stock Debian Stretch gnome-shell (with standard
extensions), mutter, dconf-settings and Xorg (no Wayland here).

This is obviously a very ugly fix that might break something,
but on preliminary testing the problem disappears and no known defects
have yet been observed.  It must slow down the UI a bit, but probably
not badly enough for anyone to notice.

Let's put this into wider testing... we must return to this later,
at least with Debian Buster.