Revert "Allow read-access to "puavoExternalId"-attribute under People.subtree"
This reverts commit 307bfeddca11b8dd67701286fc980235bb7c4eeb.

This was premature.  I thought that in password changes the ACLs to
target user were relevant, but actually it should have been more about
the ACLs to the acting user, but even that turned out to be untrue
and the relevant user is "uid=puavo,o=puavo".  But even looking up
the value of "puavoExternalId" could be avoided with some smarter coding.