Build openldap also to images, so that we get slapd that is linked with ssl instead of gnutls.  Must be built without timestamp in version string, because we need the 32-bit versions of this as well.
I do not properly remember the rationale for linking against ssl
instead of gnutls, but that is what we do on the cloud anyway,
and IIRC at least historically it had to do with stability issues
and possibly also slave synchronization problems.  The latter is
currently an issue on Debian bootservers, which I hope this will fix.