Make puavo-notify-changelog use puavo-webwindow.
This is a very ugly and error-prone way to do this. I tested
everything and while it seems to work, I still have a bad feeling
about it.

I'd love to just embed a minimal version of puavo-webwindow in this
applet, then this would be less fragile and have fewer moving parts.
But Python 2 and 3 code don't mix and right now I don't have time to
figure out how to use Webkit in Python 2 code. This applet needs to
be ported to Python 3, but that has to wait.