Run grub-installation and configuration update from puavo-image-preinst in such a way, that puavo-install-grub and grub-install scripts/programs are run in the chroot.
This is a bit trickier, but simplifies puavo-install-grub
script as --grub-basedir is no longer needed.  But the real
reason for this is that if systems differ too much (such as
updating to Debian Stretch from Ubuntu Trusty), grub-install
does not work due to missing dynamic libraries, and even otherwise
it is probably safer to use the target system environment
for updates.  On the other hand this means that puavo-install-grub
should be written so that it is understood it might be run
in a special context.  For this reason pass --hosttype parameter
to it always when doing image updates (because it can not read
the system hosttype).