Do not increase samba uid range up to 99999.
This partially reverts commit 38325a3a783ebb61ce252f4c49fa1831a3d09ddf,
as I do not fully understand what I was doing there.

The motivation for this change was that, if sambaNextRid-attribute gets
over 9999, Windows-host registrations to domain might fail (I am not
sure about this).  "sambaNextRid"-attribute apparently gets incremented
each time a new user is added to the ldap database, so this might become
a problem.  But allowing uids over 10000 may conflict with Puavo-user
uid numbers, which may be bad (not sure about this either in our current
setup).  So avoid the possible conflict, and if we actually
need to deal with the problem of Windows host registrations failing,
we should examine this a bit further.