puavomenu: Completely revamp the YAML and .desktop file loading process.
This change is too complex to be split, so here's what it does:

  - Rename several members of Program, Menu and Category classes
  - Remove some members from Program, Menu and Category classes
  - Don't build actual Program, Menu and Category objects until
    much later in the process.
  - Use generic dicts for storing YAML and .desktop file data,
    not Program, Menu or Category objects.
  - Merge YAML/.desktop data while taking care to not overwrite
    YAML data with .desktop data. This allows partial overwriting
    of .desktop files without having to specify the whole thing
    in YAML if only one string must be changed.
  - Move all loading functions to loader.py, even icon loading
  - Split the loading system into smaller functions
  - And probably several other smaller things...

I've spent the last two (almost three) days writing numerous test
cases for this code, but the test system is not ready to be
committed yet. It also has some fairly large missing things, but
it already has revealed some bugs, which are fixed in this commit.

I fear this commit will break more things than it fixes...