We want to disable flow control on ltsp* interfaces.
This is an old tweak from our ltsp-era, and may still be relevant
for netbooting clients.  We have some issues with nbd-servers not
responding (sometimes), and I suspect this fixes the issue.

Anyway, disabling flow control on eth* is wrong, because we no
longer use the bridging interface on ltsp* (by default), and
the eth* interface is renamed to ltsp*, thus ltsp* is what we
should test when disabling flow control on interfaces.  I do
not think it is necessary to disable it on other interfaces.