puavomenu: Redesigned icon cache.
The new cache is much simpler, since it does not support multiple
atlases. The old one did and multiple atlases worked, but they
were not actually used.

Another nice-to-have feature for the future is that we can remove
icons from the cache now. The removal method exists, but it's not
part of this commit. We'll need these "dynamic" icons at some
point when dynamically loaded Wine applications are added to
Puavomenu.