Take some patches from the most recent (Buster) vtun Debian-package so it can be built for Buster.
Note that we are NOT updating vtun to latest version.  This is because
vtun checks its version in its protocol, and we must support the older
vtun versions as well.  This means we can not update it on the bootserver
before we update it on access points (using puavo-wlanap), and we can
not update it on the access points before we update it on the bootserver.

There should be a solution to this, for now the solution is to use
the old vtun version (3.0.2) and then maybe get rid of it altogether.