Add "nautilus-nextcloud" and "nextcloud-desktop" packages.
This pulls "tracker" in, which we have previously not wanted
because of resource usage concerns on fatclients, but maybe
we should re-evaluate that as well... actually, it appears
"nautilus" pulls "tracker" in already, so there is no point
in fighting.  We can disable "tracker" in case it gives us
trouble, but I have a feeling ftimes have changed for the better.