puavomenu: Initial work on "dynamic" puavo-pkg installer icons.
At startup, if the program is not installed yet, the program entry
is converted into an "installer". This installer actually lauches
(or will launch) a separate program that installs the program. The
program entry (and icon) will be then dynamically updated to launch
the program.

This commit contains debugging code.