Add a new tool, "puavo-disk-clone", that allows cloning puavo-os installations (and possibly other systems on a disk) from the Puavo OS Console.  If a host has booted off from the bootserver, the system can be saved there and also restored to another host.
Booting a laptop from network mounts filesystems in puavo-os
installation, though, in which case saving/restoring does not work.

For this to work, some additional setup on the bootserver is needed,
and for now that is not going to be part of a standard bootserver
installation.  Basically setting up a "clonezilla.images"-user with
a home directory that can be accessed with ssh is enough (note that
Puavo-users can not use ssh access to bootserver).