There are some puavo-conf variables that, to be effective, should be passed on as kernel arguments.  These are currently "puavo.admin.persistent_homedirs", "puavo.boot.plymouth.theme" and "puavo.image.overlay".  In grub configuration, pass these on if they have been set in Puavo.
This improves situation in where we need to set any of these
puavo-conf variables to multiple computers in the same organisation
or school.  And even when setting host-specific configuration we
do not need to know that some key/values belong to a different place.

Some downsides to this solution: it does not help in case
these variables are passed as image-specific configuration
or as a local host-specific configuration (this could be improved
with small changes).  Also this does not fix the problem for
netbooting devices (which requires changes to bootserver).