New user registration / WiFi setup tool, version 2.0.
Initial commit. Translations are still missing (the program is a
mismash between two languages) and while it has been tested and
shown to work, it cannot be used in production yet, it needs
*A LOT* more testing.

Version 2.0 fixes all the problems in the previous version, and
adds some new features:

  - The program has been modularised better: each page is in its
    own file, for example.

  - Next/Previous buttons on most pages.

  - Far better styling/CSS usage on all pages. Page layouts have
    been thoroughly tested and tweaked.

  - The program looks (mostly) fine on all 1366x768, 1920x1080
    screens and 1600/1680 -pixel wide screens are also better
    supported now.

No Puppet rules yet on purpose. This is not meant to be used in
production yet.