Use "cdn.${topdomain}" to lookup images in Buster, replacing "images.${topdomain}".
The primary reason for this is that we need to setup a new virtualhost
so we can update certificate without breaking old systems that have not
updated to the most recent (future) version of puavo-os.

Other reason is that we distribute more than just "images" through
"images.${topdomain}": Abitti, ebs, puavo-pkg packages and rdiffs.
Thus "cdn" might be a better name.  And who knows, maybe there will
some day be a cluster of machines serving these...