Introduce "puavo.admin.certs.versions", a new puavo-conf variable (that should almost always be set only in the image).  This lists certificate versions that should be on the machine (for any past, present or future version of the system), and the first one is something that is put into use.  If the variable is empty, the latest available certificate version is used.
Right now not puavo-register, puavo-update-certificate nor
puavo-ca understand versioned certificates, but this should
also change.

The main purpose of this feature is to provide a mechanism
to move on to a new root-CA certificate chain in case of a
soon-to-arrive certificate expiration or private key
compromise.

Note that we also provide a list of root-CA certificates
in the image and use those for most operations.
This is required by diskless clients that do not have
host certificates (and thus also can not store a root-CA
certificate by themselves).