Add new "puavo-chromium-app"-script that functions as a helper for creating and starting up chromium applications. Add "Novo Desktop Streamer" and "Sphere Lite", using this new mechanism.
Unfortunately creating and updating these are a bit tricky.
Chrome/Chromium do not provide a way to startup a new profile
with these apps loaded (--app-id= does not install the app
or even if it could be made to install it, it is actually
working only on the second startup of the browser).  Thus
creating a necessary browser profile with app installed by hand
is currently a necessary step.

Running these apps should not mess up the current, actual
browser profile in any way.