Add a puavo-conf quirk for broken UEFI implementations.
Apply this for "HP ProBook 430 G2".  This is as in
https://wiki.debian.org/UEFI the section
"Force grub-efi installation to the removable media path".

So what we have is (if I understand correctly): a new shiny
technology (UEFI) to boot multiple operating systems on the same
machine.  Nicely designed.  Except that there are broken
implementations out there, that fail to do one thing: to actually
boot the operating system they are asked to boot, and then they
boot nothing.  Except that there is a workaround for that,
which is used by Microsoft Windows.  And the workaround happens
to be such, that supporting multiple operating systems on the
same machine is broken.

I am adding a quirk (like Debian does), so that we can recognize
these broken implementations.  It seems that it is *not* a good idea
to try to setup a dual-boot UEFI system on these hosts that need to set
"puavo.grub.efi.force_extra_removable" to "true" for UEFI to boot.
(One possible problem with these hosts: a Windows-update might break
booting Puavo OS, in case it overwrites the "removable media path"-EFI
file).

It is hard for me to believe engineers at HP (and other companies)
could get this wrong by accident.

WIP #521.