Bring all the changes from Buster puavo-pkgs to master (Stretch).
We want to keep only one set of puavo-pkg packages for all releases
support (currently Stretch and Buster).  Thus it makes no sense
to have Buster-specific changes in puavo-pkg packages.  Note that
not all packages should be installed (or are designed to be installable)
on both Stretch and Buster! (Puppet rules and other things decide on
this).