Fix shorewall.conf CONFIG_PATH.
The relevant part is "/" trailing "${CONFDIR}/shorewall/".
Because we are using a symbolic link from /etc/shorewall to
/state/etc/shorewall, Shorewall ignores our firewall rule
changes, unless a path is given as a parameter to "shorewall"
command line tool.  With "/" in the end, the symbolic link
is not interpreted as such, but the relevant directory
is looked up when compiling the firewall configuration.

This worked in Stretch Shorewall, so some small detail
has changed since then.

Thanks Christoph Joohs for debugging the issue.