Rewrote puavo-sync-external-files in shell, using puavo-rest-request for fetching the external files.
Added support to puavo-rest-request for multiple urls.
This makes it possible to fetch several files with only
one command (that curl *does* support), making fetching
several external files more efficient (with puavo-rest-request).

As a downside, the external file names must *not* contain
whitespace.  This is not enforced by anything (I believe)
except puavo-sync-external-files itself, if someone adds
such a name (with whitespace) to puavo-web configuration,
puavo-sync-external-files refuses to work with any
external files.  This is unfortunate and is mostly because
dealing with strings containing whitespace is trickier in
shell.  On the other hand it is good to do some input
validation to it anyway, because the external file names
must not contain path tricks, possibly overwriting some
unrelated files.

The driving reason for this change is that we want to
run puavo-sync-external-files earlier in boot (for fatclients),
making it possible to customize login screen backgrounds
through the external files mechanism.  It is better for
correctness as well, in case some feature in a system
depends on the correct external files being in place.
This implementation is somewhat faster than the ruby
version for typical files (mostly because of the faster
startup time).