Setup external files at boot time.
This includes: on fatclients fetch them from server and
run "puavo-handle-external-files-actions" so that everything
is set up properly before systemd and everything else takes over.
And also on locally booting hosts do run
"puavo-handle-external-files-actions" (not running this has been
a bug and may explain some issues occurring if a host was
booted very recently).

The unfortunate side-effect is that booting up a fatclient might
be about 0.2 seconds slower on a typical case, because it is
going to communicate with a bootserver yet one more time.