Fix apt sources.list after boot.
Fix a long-standing issue that I have not bothered to fix,
because it only affects developers and maybe some people
who try their special powers as primary user(s).  That is,
remove the local puavo-os archive from apt sources.  This
archive is broken, because we do not include it in the image
after build (because it is pointless and wastes space).

However, I have not bothered to do this previously, because
running /bin/sh + sed may take maybe 3-10 milliseconds at
boot and the problem is only for a very few.