Add support for "release-builds" target.
This make automatic something that we have done (at Opinsys)
manually for a long time.  Make one image first, then use the
build directory as a base for another, almost identical image,
called "+1"-image.  Mark these as "production images" to
images.json so we keep track of which images are important,
and then create rdiffs for the actual production image series,
plus another rdiff from the latest release image to the "+1"-image.
The purpose of the "+1"-image and the rdiffs to that image is to
have a system update test target for the latest production image,
so we can be reasonably certain that the update mechanism works
when updating to future images.