Do not remove answer files in puavo-ers-applet.
In fact, ignore all exams that do not start with "exam_".

".meb" might be an exam file, but it might also be
another kind of file, for example an answer file.
Presume that all exam files are named so that they match
a glob pattern "exam_*.meb".

Two examples of answer filenames: "answers_FI_Filosofia.meb"
and "multianswers_2020-4-7-153843-1.meb".

Closes #549.