Support Wayland as an alternative to Xorg for puavo desktop sessions.
Set "puavo.graphics.display_server" to "wayland" to test
this (default is still "xorg" of course).  Many things
are currently broken with Wayland (in puavo-os, not
commenting generally): puavomenu placement, stuff in
/etc/X11/Xsession.d does not get setup leading to some
missing functionality, display configuration management
does not work, we have not tested how the remote access
stuff should work, probably many others.  It is unlikely
we are going to move to this by default in Buster cycle,
but it should be possible to at least experiment with this.
Especially we are interested to see if some UI hangs in
Gnome regarding mouse focus and functionality work better
on the Wayland side.