Run puppet twice in apply-rules.
In the first run, accept that resource get changed, as they should
be.  On the second run, return error in case some resource
still get changed.

We used to do it this way a long time ago, but then got sloppy
and did this check manually.  No more.  Build is likely to fail,
revealing issues.