Update dash-to-panel@jderose9.github.com gnome-shell extension.
In my tests it appears the new version fixes mouse pointer grab
related issues such as this one:
https://github.com/home-sweet-gnome/dash-to-panel/issues/564
(something would grab the mouse pointer and it is impossible
to click on anything even though the mouse can be moved).

This issue was easy to reproduce with VirtualBox: make a
virtual host with eight monitors, then play with the
window preview widgets of the dash-to-panel extension.
This can also happen with other software.