Start an overhaul of the preseed installation system.
This breaks the old preseed mechanism (that has not been
used a while and is not very useful).  The idea is to lookup
preseed configurations from the installation server (bootserver),
and provide multiple options to choose from.  It should be
possible to automate mostly everything, *including* supplying
admin authentication credentials to Puavo server.