puavomenu: Improve theming support
Use the colors defined in the current theme. Now Puavomenu can switch
colors on-the-fly if the theme changes. We no longer need light or
dark versions of the colors. The CSS loader also can be simplified.

Also tweaks some of the styles and removes some useless cruft left
behind during development.