Throw away our current display management system.
It is unclear which way this will go, but we need something better.
We should try to work with Gnome rather than against it.
The current mechanism is to disable parts of Mutter to allow
us to do it ourselves.  However, what we do does not work
nicely in all situations, we only react to some events
and do not ensure that the final state is good.

The problem with Gnome display management is that it does not allow
centralized management.  A typical situation is that a host can have
multiple users, and it is not good if all they need to do something
to get sane defaults while we could just configure those somewhere.
And in case of diskless workstations the hosts themselves do not even
have persistent state to store the display configuration (except user
home directories, but that is user-specific).

The current plan is to add some configurability to affect
the Gnome display arrangement defaults (such as use "mirror"-mode
by default).  And adding to this a mechanism to catch the monitors.xml
so that it can be stored in Puavo.